LayoutTests/imported/w3c/web-platform-tests/html-aam/roles-minimum.tentative-expected.txt

Tests minimum computedrole mappings defined in HTML-AAM ([HTML-AAM Issue #454](https://github.com/w3c/html-aam/pull/454)), where the returned computed role for generic elements is expected to change based on the generic elements having attributes that would require the element no longer be generic, and potentially ignored by browsers. Most test names correspond to a unique ID defined in the spec.

Baseline div, role=generic
div with autofocus attribute.
div with explicit role=generic and autofocus attribute.
div with explicit role=none and autofocus attribute.
div with draggable attribute.
div with explicit role=generic and draggable attribute.
div with explicit role=none and draggable attribute.
Show popover div
div with draggable attribute and errant role attribute value.
unnamed section with draggable attribute.
div with explicit role=article and draggable attribute.
article element with draggable attribute.
named section with draggable attribute.
cite element with draggable attribute.

PASS el-div
FAIL el-div-autofocus-attr assert_equals: <div data-testname="el-div-autofocus-attr" data-expectedrole="group" autofocus="" class="ex">div with autofocus attribute.</div> expected "group" but got "generic"
FAIL el-div-autofocus-attr-role-generic assert_equals: <div data-testname="el-div-autofocus-attr-role-generic" data-expectedrole="group" autofocus="" role="generic" class="ex">div with explicit role=generic and autofocus attribute.</div> expected "group" but got "generic"
FAIL el-div-autofocus-attr-role-none assert_equals: <div data-testname="el-div-autofocus-attr-role-none" data-expectedrole="group" autofocus="" role="none" class="ex">div with explicit role=none and autofocus attribute.</div> expected "group" but got "none"
FAIL el-div-draggable-attr assert_equals: <div data-testname="el-div-draggable-attr" data-expectedrole="group" draggable="" class="ex">div with draggable attribute.</div> expected "group" but got "generic"
FAIL el-div-draggable-attr-role-generic assert_equals: <div data-testname="el-div-draggable-attr-role-generic" data-expectedrole="group" draggable="" role="generic" class="ex">div with explicit role=generic and draggable attribute.</div> expected "group" but got "generic"
FAIL el-div-draggable-attr-role-none assert_equals: <div data-testname="el-div-draggable-attr-role-none" data-expectedrole="group" draggable="" role="none" class="ex">div with explicit role=none and draggable attribute.</div> expected "group" but got "generic"
FAIL el-div-popover-attr assert_equals: <div data-testname="el-div-popover-attr" id="pd" data-expectedrole="group" popover="" class="ex">div with popover attribute, role=group</div> expected "group" but got "generic"
FAIL el-div-draggable-attr-invalid-role assert_equals: <div data-testname="el-div-draggable-attr-invalid-role" data-expectedrole="group" draggable="" role="foo" class="ex">div with draggable attribute and errant role attribute value.</div> expected "group" but got "generic"
FAIL el-section-draggable-attr assert_equals: <section data-testname="el-section-draggable-attr" data-expectedrole="group" draggable="" class="ex">unnamed section with draggable attribute.</section> expected "group" but got ""
PASS el-div-draggable-attr-role-article
PASS el-article-draggable-attr
PASS el-section-named-draggable-attr
FAIL el-cite-draggable-attr assert_equals: <cite data-testname="el-cite-draggable-attr" data-expectedrole="html-cite" draggable="" class="ex">cite element with draggable attribute.</cite> expected "html-cite" but got ""
